Fix cudf 25.08 build
* cudf 25.08 in Velox requires C++20 due to the use of requires clauses.
To fix, conditional C++20 is enabled for cudf modules only, while the
Velox core remains on C++17.

* Using C++20 for cudf with C++17 Velox headers results in deprecation
warnings. The -Wno-deprecated flag is added to cudf targets only.

* The parquet_chunked_writer is deprecated and replaced by
chunked_parquet_writer. The code is updated to use the non-deprecated
writer or to ignore deprecated warnings as needed.

* The join operators are refactored to cudf/join/, which is a breaking
change and requires updates in the Velox code.

* cudf 25.08 is bundled with CCCL 3.0.0, which itself includes CUB and
Thrust. This can conflict with the versions supplied by CUDA on the
system. In cases where CUB is required or linked but Thrust is not,
compilation errors can occur because Thrust defaults to the version
bundled with CCCL. To resolve this, whenever CUB is required or linked,
Thrust must also be required or linked to ensure the same version is
used for both and to avoid mixing versions.